server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
openai>=1.30.0
numpy>=1.24.0
python-dotenv>=1.0.0
transformers>=4.37.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
